refactor: extract frontend core modules
