+++ frontend/src/features/lms/components/AppSidebarSectionsCore.tsx
@@ -1,0 +1,167 @@
+import { isNavItemActive, roleLabel } from '../config';
+import type { LmsNavKey, LmsPageId, SidebarDock, ThemeMode } from '../types';
+
+type NavGroup = {
+  label: string;
+  items: Array<{
+    page: LmsPageId;
+    label: string;
+    icon: string;
+  }>;
+};
+
+type AppSidebarHeaderSectionProps = {
+  collapsed: boolean;
+  onHome: () => void;
+  onCloseMobile: () => void;
+  onToggleCollapsed: () => void;
+};
+
+type AppSidebarNavSectionProps = {
+  groups: NavGroup[];
+  activePage: LmsPageId;
+  activeNavKey: LmsNavKey;
+  collapsed: boolean;
+  dock: SidebarDock;
+  onNavigate: (page: LmsPageId) => void;
+};
+
+type AppSidebarFooterSectionProps = {
+  avatarTone: string;
+  collapsed: boolean;
+  sessionName: string;
+  role: LoginRole;
+  dock: SidebarDock;
+  theme: ThemeMode;
+  onLogout: () => void;
+  onToggleDock: () => void;
+  onToggleTheme: () => void;
+};
+
+type LoginRole = 'ADMIN' | 'INSTRUCTOR' | 'STUDENT';
+
+export function AppSidebarHeaderSection({ collapsed, onHome, onCloseMobile, onToggleCollapsed }: AppSidebarHeaderSectionProps) {
+  return (
+    <div className="flex h-16 items-center justify-between gap-2 border-b border-[var(--app-border)] px-4">
+      <button
+        type="button"
+        onClick={() => {
+          onHome();
+          onCloseMobile();
+        }}
+        className={`flex items-center gap-2.5 ${collapsed ? 'mx-auto' : ''}`}
+        title="메인 화면으로 이동"
+      >
+        <div className="flex h-9 w-9 items-center justify-center rounded-xl bg-gradient-to-br from-indigo-500 to-violet-600 text-white shadow-md shadow-indigo-500/25">
+          <i className="ri-play-circle-fill text-[18px]" />
+        </div>
+        {!collapsed ? <span className="text-[15px] font-extrabold tracking-tight text-[var(--app-text)]">내맘대로</span> : null}
+      </button>
+      <button
+        type="button"
+        onClick={onToggleCollapsed}
+        className="hidden h-8 w-8 items-center justify-center rounded-lg text-[var(--app-text-muted)] transition hover:bg-[var(--app-surface-soft)] hover:text-[var(--app-text)] lg:flex"
+        title={collapsed ? '사이드바 펼치기' : '사이드바 접기'}
+      >
+        <i className={collapsed ? 'ri-arrow-right-s-line text-[17px]' : 'ri-arrow-left-s-line text-[17px]'} />
+      </button>
+      <button
+        type="button"
+        onClick={onCloseMobile}
+        className="flex h-8 w-8 items-center justify-center rounded-lg text-[var(--app-text-muted)] transition hover:bg-[var(--app-surface-soft)] hover:text-[var(--app-text)] lg:hidden"
+        title="메뉴 닫기"
+      >
+        <i className="ri-close-line text-lg" />
+      </button>
+    </div>
+  );
+}
+
+export function AppSidebarNavSection({ groups, activePage, activeNavKey, collapsed, dock, onNavigate }: AppSidebarNavSectionProps) {
+  return (
+    <nav className={`flex-1 overflow-y-auto ${collapsed ? 'px-2' : 'px-3'} py-4`}>
+      {groups.map((group) => (
+        <div key={group.label} className="mb-5">
+          {!collapsed ? <div className="mb-2 px-3 text-[10px] font-bold uppercase tracking-[0.12em] text-[var(--app-text-muted)]">{group.label}</div> : null}
+          <div className="space-y-0.5">
+            {group.items.map((item) => {
+              const active = isNavItemActive(item, activeNavKey) || item.page === activePage;
+              return (
+                <button
+                  key={item.page}
+                  type="button"
+                  onClick={() => onNavigate(item.page)}
+                  title={collapsed ? item.label : undefined}
+                  className={`relative flex w-full items-center gap-2.5 rounded-xl px-3 py-2.5 text-left text-[13px] font-medium transition-all duration-150 ${
+                    active
+                      ? 'bg-[var(--app-accent-soft)] text-[var(--app-accent)] shadow-sm'
+                      : 'text-[var(--app-text-secondary)] hover:bg-[var(--app-surface-soft)] hover:text-[var(--app-text)]'
+                  }`}
+                >
+                  {active ? <span className={`absolute ${dock === 'right' ? 'right-0 rounded-l-full' : 'left-0 rounded-r-full'} top-1/2 h-5 w-[3px] -translate-y-1/2 bg-[var(--app-accent)]`} /> : null}
+                  <i className={`${item.icon} w-[22px] text-center text-[17px] ${active ? '' : 'opacity-70'}`} />
+                  {!collapsed ? <span>{item.label}</span> : null}
+                </button>
+              );
+            })}
+          </div>
+        </div>
+      ))}
+    </nav>
+  );
+}
+
+export function AppSidebarFooterSection({
+  avatarTone,
+  collapsed,
+  sessionName,
+  role,
+  dock,
+  theme,
+  onLogout,
+  onToggleDock,
+  onToggleTheme,
+}: AppSidebarFooterSectionProps) {
+  return (
+    <div className="border-t border-[var(--app-border)] px-3 py-3">
+      <div className={`flex items-center gap-2.5 rounded-xl px-3 py-2.5 ${collapsed ? 'justify-center' : ''}`}>
+        <div className={`flex h-9 w-9 items-center justify-center rounded-xl text-[13px] font-bold text-white shadow-sm ${avatarTone}`}>{sessionName[0]}</div>
+        {!collapsed ? (
+          <div className="min-w-0 flex-1">
+            <div className="truncate text-[13px] font-semibold text-[var(--app-text)]">{sessionName}</div>
+            <div className="text-[11px] text-[var(--app-text-muted)]">{roleLabel(role)}</div>
+          </div>
+        ) : null}
+        {!collapsed ? (
+          <button
+            type="button"
+            onClick={onLogout}
+            className="flex h-8 w-8 items-center justify-center rounded-lg text-[var(--app-text-muted)] transition hover:bg-[var(--app-surface-soft)] hover:text-red-500"
+            title="로그아웃"
+          >
+            <i className="ri-logout-box-r-line text-[16px]" />
+          </button>
+        ) : null}
+      </div>
+      {!collapsed ? (
+        <div className="mt-2.5 flex items-center gap-2 px-1">
+          <button
+            type="button"
+            onClick={onToggleDock}
+            className="flex-1 rounded-xl border border-[var(--app-border)] bg-[var(--app-surface-soft)] px-3 py-2 text-[11px] font-semibold text-[var(--app-text-secondary)] transition hover:bg-[var(--app-surface-hover)]"
+          >
+            {dock === 'left' ? '오른쪽 고정' : '왼쪽 고정'}
+          </button>
+          <button
+            type="button"
+            onClick={onToggleTheme}
+            className="flex h-9 w-9 items-center justify-center rounded-xl border border-[var(--app-border)] bg-[var(--app-surface-soft)] text-[var(--app-text-muted)] transition hover:bg-[var(--app-surface-hover)] hover:text-[var(--app-text)]"
+            title={theme === 'light' ? '다크 모드' : '라이트 모드'}
+          >
+            <i className={theme === 'light' ? 'ri-moon-line text-[15px]' : 'ri-sun-line text-[15px]'} />
+          </button>
+        </div>
+      ) : null}
+    </div>
+  );
+}
