+++ frontend/src/styles-animations.css
@@ -1,220 +1,2 @@
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
+@import './styles-animations-core.css';
 
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@keyframes fadeInUp {
-  from {
-    opacity: 0;
-    transform: translateY(20px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@keyframes slideInLeft {
-  from {
-    opacity: 0;
-    transform: translateX(-18px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-@keyframes slideInRight {
-  from {
-    opacity: 0;
-    transform: translateX(18px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-@keyframes scaleIn {
-  from {
-    opacity: 0;
-    transform: scale(0.94);
-  }
-
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-@keyframes shimmer {
-  0% {
-    background-position: -200% 0;
-  }
-
-  100% {
-    background-position: 200% 0;
-  }
-}
-
-@keyframes pulse-soft {
-  0%,
-  100% {
-    opacity: 1;
-  }
-
-  50% {
-    opacity: 0.55;
-  }
-}
-
-@keyframes float {
-  0%,
-  100% {
-    transform: translateY(0);
-  }
-
-  50% {
-    transform: translateY(-8px);
-  }
-}
-
-@keyframes glow {
-  0%,
-  100% {
-    box-shadow: 0 0 20px rgba(99, 102, 241, 0.15);
-  }
-
-  50% {
-    box-shadow: 0 0 40px rgba(99, 102, 241, 0.25);
-  }
-}
-
-@keyframes gradient-shift {
-  0% {
-    background-position: 0% 50%;
-  }
-
-  50% {
-    background-position: 100% 50%;
-  }
-
-  100% {
-    background-position: 0% 50%;
-  }
-}
-
-@keyframes slideDown {
-  from {
-    opacity: 0;
-    transform: translateY(-8px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@keyframes progress-fill {
-  from {
-    width: 0;
-  }
-}
-
-.animate-fade-in {
-  animation: fadeIn 0.4s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-.animate-fade-in-up {
-  animation: fadeInUp 0.5s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-.animate-slide-left {
-  animation: slideInLeft 0.4s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-.animate-slide-right {
-  animation: slideInRight 0.4s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-.animate-scale-in {
-  animation: scaleIn 0.35s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-.animate-shimmer {
-  background: linear-gradient(90deg, transparent 0%, rgba(255, 255, 255, 0.06) 50%, transparent 100%);
-  background-size: 200% 100%;
-  animation: shimmer 2.5s ease-in-out infinite;
-}
-
-.animate-pulse-soft {
-  animation: pulse-soft 2s ease-in-out infinite;
-}
-
-.animate-float {
-  animation: float 4s ease-in-out infinite;
-}
-
-.animate-glow {
-  animation: glow 3s ease-in-out infinite;
-}
-
-.animate-slide-down {
-  animation: slideDown 0.3s ease-out both;
-}
-
-.animate-progress {
-  animation: progress-fill 0.8s ease-out both;
-}
-
-.stagger > * {
-  animation: fadeIn 0.4s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-.stagger > *:nth-child(1) {
-  animation-delay: 0.03s;
-}
-
-.stagger > *:nth-child(2) {
-  animation-delay: 0.07s;
-}
-
-.stagger > *:nth-child(3) {
-  animation-delay: 0.11s;
-}
-
-.stagger > *:nth-child(4) {
-  animation-delay: 0.15s;
-}
-
-.stagger > *:nth-child(5) {
-  animation-delay: 0.19s;
-}
-
-.stagger > *:nth-child(6) {
-  animation-delay: 0.23s;
-}
-
-.stagger > *:nth-child(7) {
-  animation-delay: 0.27s;
-}
-
-.stagger > *:nth-child(8) {
-  animation-delay: 0.31s;
-}
-
-.stagger > *:nth-child(9) {
-  animation-delay: 0.35s;
-}
